Package .loxone.core moved to .loxone.internal.core
