+++ addons/binding/org.openhab.binding.loxone/src/main/java/org/openhab/binding/loxone/handler/LoxoneMiniserverHandler.java
@@ -45,21 +45,21 @@ import org.eclipse.smarthome.core.types.RefreshType;
 import org.eclipse.smarthome.core.types.StateDescription;
 import org.eclipse.smarthome.core.types.StateOption;
 import org.openhab.binding.loxone.config.LoxoneMiniserverConfig;
-import org.openhab.binding.loxone.core.LxCategory;
-import org.openhab.binding.loxone.core.LxContainer;
-import org.openhab.binding.loxone.core.LxControl;
-import org.openhab.binding.loxone.core.LxControlInfoOnlyAnalog;
-import org.openhab.binding.loxone.core.LxControlInfoOnlyDigital;
-import org.openhab.binding.loxone.core.LxControlJalousie;
-import org.openhab.binding.loxone.core.LxControlLightController;
-import org.openhab.binding.loxone.core.LxControlPushbutton;
-import org.openhab.binding.loxone.core.LxControlRadio;
-import org.openhab.binding.loxone.core.LxControlSwitch;
-import org.openhab.binding.loxone.core.LxControlTextState;
-import org.openhab.binding.loxone.core.LxOfflineReason;
-import org.openhab.binding.loxone.core.LxServer;
-import org.openhab.binding.loxone.core.LxServerListener;
 import org.openhab.binding.loxone.internal.LoxoneChannelTypeProvider;
+import org.openhab.binding.loxone.internal.core.LxCategory;
+import org.openhab.binding.loxone.internal.core.LxContainer;
+import org.openhab.binding.loxone.internal.core.LxControl;
+import org.openhab.binding.loxone.internal.core.LxControlInfoOnlyAnalog;
+import org.openhab.binding.loxone.internal.core.LxControlInfoOnlyDigital;
+import org.openhab.binding.loxone.internal.core.LxControlJalousie;
+import org.openhab.binding.loxone.internal.core.LxControlLightController;
+import org.openhab.binding.loxone.internal.core.LxControlPushbutton;
+import org.openhab.binding.loxone.internal.core.LxControlRadio;
+import org.openhab.binding.loxone.internal.core.LxControlSwitch;
+import org.openhab.binding.loxone.internal.core.LxControlTextState;
+import org.openhab.binding.loxone.internal.core.LxOfflineReason;
+import org.openhab.binding.loxone.internal.core.LxServer;
+import org.openhab.binding.loxone.internal.core.LxServerListener;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
